move tests from pos to run
