+++ test/tasty/neg/src-2/TestHello.check
@@ -1,33 +1,33 @@
-TestHello_fail.scala:6: error: inferred type arguments [List[String]] do not conform to method higherBounded2's type parameter bounds [T <: List[_ <: Int]]
+TestHello_fail.scala:4: error: inferred type arguments [List[String]] do not conform to method higherBounded2's type parameter bounds [T <: List[_ <: Int]]
   HelloWorld.higherBounded2(List(""))
              ^
-TestHello_fail.scala:6: error: type mismatch;
+TestHello_fail.scala:4: error: type mismatch;
  found   : List[String]
  required: T
   HelloWorld.higherBounded2(List(""))
                                 ^
-TestHello_fail.scala:7: error: inferred type arguments [List[List[String]]] do not conform to method higherBounded3's type parameter bounds [T <: List[List[_ <: Int]]]
+TestHello_fail.scala:5: error: inferred type arguments [List[List[String]]] do not conform to method higherBounded3's type parameter bounds [T <: List[List[_ <: Int]]]
   HelloWorld.higherBounded3(List(List("")))
              ^
-TestHello_fail.scala:7: error: type mismatch;
+TestHello_fail.scala:5: error: type mismatch;
  found   : List[List[String]]
  required: T
   HelloWorld.higherBounded3(List(List("")))
                                 ^
-TestHello_fail.scala:8: error: inferred type arguments [scala.util.Left[String,String]] do not conform to method higherBounded4's type parameter bounds [T <: Either[_ <: Int, String]]
+TestHello_fail.scala:6: error: inferred type arguments [scala.util.Left[String,String]] do not conform to method higherBounded4's type parameter bounds [T <: Either[_ <: Int, String]]
   HelloWorld.higherBounded4(Left[String,String](""))
              ^
-TestHello_fail.scala:8: error: type mismatch;
+TestHello_fail.scala:6: error: type mismatch;
  found   : scala.util.Left[String,String]
  required: T
   HelloWorld.higherBounded4(Left[String,String](""))
                                                ^
-TestHello_fail.scala:12: error: kinds of the type arguments (helloworld.TestHello.Show) do not conform to the expected kinds of the type parameters (type F).
+TestHello_fail.scala:10: error: kinds of the type arguments (helloworld.TestHello.Show) do not conform to the expected kinds of the type parameters (type F).
 helloworld.TestHello.Show's type parameters do not match type F's expected parameters:
 type A is contravariant, but type _ is declared covariant
   HelloWorld.higherBounded5[Show]
                            ^
-TestHello_fail.scala:13: error: kinds of the type arguments (List) do not conform to the expected kinds of the type parameters (type F).
+TestHello_fail.scala:11: error: kinds of the type arguments (List) do not conform to the expected kinds of the type parameters (type F).
 List's type parameters do not match type F's expected parameters:
 type A is covariant, but type _ is declared contravariant
   HelloWorld.higherBounded6[List]
