+++ test/tasty/neg/src-2/TestHello_fail.scala
@@ -1,8 +1,6 @@
 package helloworld
 
 object TestHello {
-  // val helloworld = HelloWorld.msg
-  // HelloWorld.acceptsOnlyMsg1(helloworld)
   HelloWorld.higherBounded2(List(""))
   HelloWorld.higherBounded3(List(List("")))
   HelloWorld.higherBounded4(Left[String,String](""))
