--- test/tasty/pos/src-2/tastytest/TestHello.scala
@@ -1,7 +1,0 @@
-package tastytest
-
-object TestHello extends Suite("TestHello") {
-
-  test(assert(HelloWorld.`1) my long test assertion that 1%3 == 3` === 3)) // move to run after 0.21.0-RC1 for binary compat
-
-}
