--- test/tasty/pos/src-2/tastytest/TestNewJVMEnvLive.scala
@@ -1,17 +1,0 @@
-package tastytest
-
-import printing._
-
-object NewJVMEnv {
-
-  final class TestClass {
-    override def hashCode(): Int = 123
-  }
-
-  val JVMEnvLive: JVMEnvLive = new JVMEnvLive {}
-  assert(
-    (JVMEnvLive)
-      .evalInPrinter(Printer.ObjectToString)(new TestClass) == "tastytest.TestPrinter$TestClass@123"
-  )
-
-}
