--- test/tasty/pos/src-3/tastytest/HelloWorld.scala
@@ -1,5 +1,0 @@
-package tastytest
-
-object HelloWorld {
-  def `1) my long test assertion that 1%3 == 3`: Int = 1%3
-}
