+++ test/tasty/run/src-2/tastytest/TestHello.scala
@@ -24,6 +24,7 @@ object TestHello extends Suite("TestHello") {
   test(assert(HelloWorld.func1(33) === 33))
   test(assert((HelloWorld.lzy: "lazy") === "lazy"))
   test(assert(HelloWorld.acceptsOnlyMsg4(HelloWorld.msg4) === "Hello, World!Hello, World!"))
+  test(assert(HelloWorld.`1) my long test assertion that 1^3 == 2` === 2))
   // test(assert(HelloWorld.`<init>` === 157)) // wait until https://github.com/lampepfl/dotty/issues/7799
 
   type OrInt[A] = Either[Int, A]
