+++ test/tasty/run/src-3/tastytest/HelloWorld.scala
@@ -26,6 +26,7 @@ object HelloWorld {
   def func1[A]: A => A = x => x
   def acceptsOnlyMsg4(m: msg4.type): String = m + m
   final lazy val lzy = "lazy"
+  def `1) my long test assertion that 1^3 == 2`: Int = 1^3
   // def `<init>`: Int = 157 // broken in https://github.com/lampepfl/dotty/issues/7799
 
 }
